refactor: extract analytics cards and modularize budget/import sheets
Break down large screens and sheets into focused, reusable pieces:

- Analytics: extract MoneyPulse, PriorityActionPlan, SpendingHotspots,
  UpcomingBills, and SectionCard components from analytics.tsx
- BudgetFormSheet: split out category/option picker sheets and a
  useBudgetPreview hook
- ImportCSVSheet: extract ImportIntro and ImportProgressCard plus a
  useCsvImport feature hook
- create-transaction: extract useTransactionFormSheets hook
- Clarify the Burnt icon @ts-expect-error and bump RN version note in AGENTS.md
