+++ components/analytics/SectionCard.tsx
@@ -1,0 +1,16 @@
+import { styled, View } from 'tamagui';
+
+export const SectionCard = styled(View, {
+  backgroundColor: '$gray3',
+  borderRadius: '$4',
+  padding: '$4',
+  marginBottom: '$3',
+});
+
+export const MetricTile = styled(View, {
+  backgroundColor: '$gray4',
+  borderRadius: '$3',
+  padding: '$3',
+  borderWidth: 1,
+  borderColor: '$gray5',
+});
